Analyzer: Pass start/end times correctly
When one table wanted information from another table, it used
get_sqobject to get the basic object on which to operate. But in the
process, it failed to provide additional context of start and end times
that the user had qualified for the analysis.

This patch adds an intermediate function to specifically avoid this
mistake within the engines/pandas folder.