Interfaces: Fix assert to handle subinterfaces and Junos
Interface assert wasn't working properly in a bunch of scenarios, and
these got missed from being caught because of an incorrectly placed
dropna. This patch now handles the following two cases correctly:
* Junos device interfaces peering with non-Junos devices
* routed subinterfaces

In addition, we also provide an option to ignore missing peers either
due to the reason that the device is unpolled or is just missing (
say because of no LLDP info for an interface). However, we can't simply
flag all interfaces as not having a peer since most server interfaces do
not enable LLDP, and so server-facing ports have no peer. We try to
ignore such interfaces from being flagged as having no peers.