sqobject: Handle singular/plural confusion
Unfortunately, we've introduced a bit of a confusion in our output
and code. REST, the CLI all assume the singular form while the rest
of the code assumes the plural form. This leads to user confusion as
to the choice of singular or plural. This is most common with the
following tables: routes, macs, interfaces, tables. For these, just
make the correction to the right name whichever form its presented in,
singular or plural.