Poller: Add support for having multiple command versions
Many NOS have either different commands for the same output OR
different outputs for the same command OR both, across different
versions of the NOS. For example, Junos has "show lldp neighbors"
and "show lldp neighbors detail" to be able to retrieve the peering
interface name.

This adds support for this feature.