Add schema as a property of the SqPandasEngine
When we need to determine the display fields or fetch some other
info from the schema, the existing property called schemas which was
really the list of all schemas had to be referenced to extract the
schema appropriate for the table. This is cumbersome and duplicates
the logic in multiple places.

Instead, we renamed the existing property as all_schemas and added
a new property called schema to access the SchemaTable for the table
easily.