Network find: redux
This patch addresses doing network find correctly when start and end
times are provided. Instead of merely running queries across all tables
in aggregate, we need to process each row from the ARP table separately.
This is because we need to match up with this time to identify the
correct outgoing interface including tracing the path through an L2
network to the final L2 switch.

Furthermore, many users suggested we display both bond members and the
bond interface itself instead of using resolve_bond=True. That is also
addressed in this patch.

We're making network find a pure find attach point locator for now.
Its only used to find the first hop switch that an IP/MAC is attached
to.

We'll expand to include others via either other verbs or some other
modification to find.

Finally, query-str was not handled correctly for network find even if
the user specified one. Fixed that too.