SqPlugin: Make get_plugins work without any additional params for a plugin
SqPlugin has a class method called get_plugins that returns all the
plugins associated with a plugin type. However this required an additional
parameter called search_pkg which was not really necessary. It can be
derived automatically from the module name. This patch does that.

Additionally, the name of the plugin returned when the plugin was in
a subdir as opposed to the same dir was not correct. This patch fixes
that as well.

The method signature isn't changed, and so if a caller wants to specify
the search path that still works

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>