SqPlugin: Make get_plugins work without any additional params for a p… (#500)
* SqPlugin: Make get_plugins work without any additional params for a plugin

SqPlugin has a class method called get_plugins that returns all the
plugins associated with a plugin type. However this required an additional
parameter called search_pkg which was not really necessary. It can be
derived automatically from the module name. This patch does that.

Additionally, the name of the plugin returned when the plugin was in
a subdir as opposed to the same dir was not correct. This patch fixes
that as well.

The method signature isn't changed, and so if a caller wants to specify
the search path that still works

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Add .pylintrc to make pylint usable
More updates to pylintrc to make messages manageable

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
(cherry picked from commit 98458c41fc55653ac1bf15b5f24d5c6120f206d9)

* sq_plugin: More cleanup and changing code to handle new plugin model

transitive keyward didn't really work, and no plugin is using it so we
removed it

Make existing code fit into new SqPlugin model

SqDB, SqEngineObj, SqObjects and more have been switched to using
the standard SqPlugin model with the get_plugins() routing being
used consistently throughout the code. The poller is one area that
still needs to adapt to this model

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Add tests to verify new plugin

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>