Add support for REST engine
The REST engine merely allows the CLI/GUI/Python API to communicate
with a REST server to access data that is not local to where these
front ends (CLI/GUI/Python API) are run. The assumption is that the
REST engine translates the user commands into REST API calls and
then returns the data as a pandas DF just like the other engines.

Besides this engine add, this patch also adds options to enable
the specification of the REST engine, and removing any hardcoded
value for 'pandas' as the only engine.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>