Add new context dataclass
A suzieq context is passed between the backend and the frontend. This
context was duplicated inside the CLI and the sqobject library. This
patch pulls out the context into a separate file called SqContext (no
change in the name), and makes SqContext a dataclass. This context is
shared between the python API and the CLI.

The rest of the changes are to handle this modification, and some of
the simpler pylint fixes.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>