CLI: Setup framework for allowing filters to apply to all commands
Before this patch, filters were differently applied to different
commands(verbs) of an object (such as route, interface etc.). This
is often an impediment to effective querying of the data by the
users.

This patch allows all commands to take the same set of filters for
all methods (except describe and help) of an object/service/table. This
required patching a nubia library that complained of duplicate arguments
even though the arguments weren't really duplicates. This is supplied
by the new file nubia_patch.

This approach also allowed us to dramatically simplify the code by
moving most, if not all, of the code associated with an object into a
common function for each verb in command.py. This also helps makes
errors and other behaviors consistent across all commands.

The remaining verb that needs massaging is aver/assert, but that needs
a parameter to be renamed in either device/sqPoller or as passed to
aver.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>