Ensure device key generated in inventory includes port
When the poller controller generates the inventory files for a worker
poller, the current code encodes each entry with only the namespace and
device address. In certain cases, the device address maybe the same, but
the ports will be different. Fix this error by having the key of each
entry in the inventory include the port as well.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>