Normalized VLAN service for IOSXE
On IOS/IOSXE platforms, the command 'show vlan' only shows access vlan
ports, not the trunk ports. This is unlike other NOS such as even NXOS.
To get information about the trunking ports, you have to get the output
of the command 'show interface trunk'.

This patch normalizes VLAN service support for IOS/XE by merging the
info from these two commands so that the output is similar to what it is
like for other NOS. Furthermore, at this point, we only store the VLANs
associated with an interface that are in STP forwarded and unpruned
state, not the configured VLANs or active in mgmt domain VLANs.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>