Simnode upd8 (#527)
* Simplify simnode setup

Using simnode as a way to suck data into suzieq can be used outside of
tests. It feels like a valuable use case all of its own. So, we can
simplify the code substantially. The requirement for NOS & version are
unnecessary given that the poller decides on all that via auto-discovery
anyway.

In this patch, we simplify the use of simnode to:
   * Just take a pointer to the top-level dir under which there are
     multiple host dirs, each containing the command output from that
     host.
   * if you want to specify only certain hosts that need to be polled,
     then use the devices file specification approach
   * Optional port and transport can be provided.

If you have a dir with mixed hosts, then you have to rely on specifying
the protocol to be used with each host via the devices file separately.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* simnode: Move the scripts into suzieq/utilities as they serve more than tests

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* simnode: pylint fixes, mv script outside tests

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* fix off-by-1 in port number assignment

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>