Bunch of patches to better support IOSXE platform (#526)
* IOSXE MAC parsing: ifnames can be anything but spaces

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOXE: fix a few parsing errors

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Fix parsing tests to handle some IOSXE cases

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Ansible source: qualify a device by its address AND port

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Poller: Handle non-std chars in dirnames during exec of coalescer

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Handle the various forms of IOSXE version string

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: More fixes to interface parsing

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Add tunnel as an additionl type for IOSXE,
Support for checking SVIs do not have blank MACs

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Validate that we have mbr info for all bonds

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: Handle Cat4500 MAC table format

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: Handle Cat4K MAC address table format

Cat4k has a different MAC address table compared to other IOSXE
platforms. There's an additional protocol field, plus there are
multicast entries called out separately. Furthermore, for multicast
entries, the outgoing interface list has more than one interface and we
need to expand all of the entries.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: Show CDP parsing fixes to handle Cat4k

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: More interface renaming madness

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Interfaces: initialize dataframe with empty array correctly

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Normalized VLAN service for IOSXE

On IOS/IOSXE platforms, the command 'show vlan' only shows access vlan
ports, not the trunk ports. This is unlike other NOS such as even NXOS.
To get information about the trunking ports, you have to get the output
of the command 'show interface trunk'.

This patch normalizes VLAN service support for IOS/XE by merging the
info from these two commands so that the output is similar to what it is
like for other NOS. Furthermore, at this point, we only store the VLANs
associated with an interface that are in STP forwarded and unpruned
state, not the configured VLANs or active in mgmt domain VLANs.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* pylint fixes

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Simplify column init with empty array

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* IOSXE: make VLAN 0 for .0 subinterfaces

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Cumulus: interface parsing error fix

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* pylint fixes

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>