GUI: Tons of bug fixes, feature fixes and cleanups
* Make row style come from a dict instead of lots of if-else
* Handle empty dataframes
* Disable selecting devconfig from list of tables
* GUI Xplore: Add experimental button
* Set network as default table in xplore
* Fix logo stretching with smaller windows
* Add warning message in path page
* Provide option to start GUI port to be something other than the
  default 8501, start on port 8501 otherwise. Print startup message,
  pylint fixes. Redirect stdout, stderr

And finally, a big whopper:
We always create a URL to capture a page's widget state. However, while
you're on a page and changing things, its the state that's saved that
counts, not the URL because of the way callbacks work in Streamlit:
first the callback is invoked and then the entire app is re-run. This
means however that the URL of the page will be rendered to the previous
successful render rather than the current values of widgets because we
were not saving the page URL in the callback. This resulted in several
bizzare behaviors.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
Signed-off-by: LucaNicosia <lucassa5@live.com>
Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>
Signed-off-by: Claudio Usai <c.usai186@gmail.com>