Network find: Remove artificial time window around lookup
Network find applied a +/- 30 second time window around the MAC
table lookup. The time window was around the timestamp provided
by the ARPND table lookup of the IP address. The idea of the time
window was to avoid the issue of a MAC table entry that might've
aged out while the ARP entry was still intact. But this creates a
problem if on the contrary the MAC table entry came in a few minutes
after the ARP table entry.

So, this patch:
* Honors user-provided start and end times
* Removes any artificial time window added around MAC table lookup
* Adds a column called l2miss which will be True if the MAC was
  not found in the MAC table, for a given ARPND entry.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>