Andrea fixes (#561)
* Handle errors in address format for network find

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* GUI: network find address needs to be passed as a list

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* ARPND summarize was broken when only a selected DF was used

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* REST: network find address is a list, not str

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* fixup to network.py

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* GUI: Tons of bug fixes, feature fixes and cleanups

* Make row style come from a dict instead of lots of if-else
* Handle empty dataframes
* Disable selecting devconfig from list of tables
* GUI Xplore: Add experimental button
* Set network as default table in xplore
* Fix logo stretching with smaller windows
* Add warning message in path page
* Provide option to start GUI port to be something other than the
  default 8501, start on port 8501 otherwise. Print startup message,
  pylint fixes. Redirect stdout, stderr

And finally, a big whopper:
We always create a URL to capture a page's widget state. However, while
you're on a page and changing things, its the state that's saved that
counts, not the URL because of the way callbacks work in Streamlit:
first the callback is invoked and then the entire app is re-run. This
means however that the URL of the page will be rendered to the previous
successful render rather than the current values of widgets because we
were not saving the page URL in the callback. This resulted in several
bizzare behaviors.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
Signed-off-by: LucaNicosia <lucassa5@live.com>
Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>
Signed-off-by: Claudio Usai <c.usai186@gmail.com>

* Network find: Remove artificial time window around lookup

Network find applied a +/- 30 second time window around the MAC
table lookup. The time window was around the timestamp provided
by the ARPND table lookup of the IP address. The idea of the time
window was to avoid the issue of a MAC table entry that might've
aged out while the ARP entry was still intact. But this creates a
problem if on the contrary the MAC table entry came in a few minutes
after the ARP table entry.

So, this patch:
* Honors user-provided start and end times
* Removes any artificial time window added around MAC table lookup
* Adds a column called l2miss which will be True if the MAC was
  not found in the MAC table, for a given ARPND entry.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* update network tests

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Address: Fix VRF field

The VRF field was not correctly set for bridged and port-channel ports.
This patch fixes that and also fixes the tests to capture the correct
output.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Update pillow due to security, streamlit to 1.3.0

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>