Pandas: Support nested augmented columns
An augmented column is a virtual column constructed by computing
data on a bunch of other columns from the same table. However, the
implementation didn't support nested augmented columns i.e. an
augmented column relying on other augmented columns. This patch
fixes that.

vlanList augmented column in interfaces is the first to use this.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>