Interfaces: Support portmode, pvid and vlanList per interface
In L2 environments, portmode, port vlan id (pvid) and allowed VLAN list
are important characteristics associated with an interface. This patch
adds support for querying and displaying this information.

This patch also adds enhancement to interface assert to verify that the
two ends of a link have the same portmode and pvid.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>