Routes: Get rid of populating oifs
Not every NOS provides the outgoing interface associated with a nexthop
for a given route. We attempted to normalize this by trying to fill in
the NOS by doing recursive route lookups. At scale, this blows up very
quickly.

The main reason for adding the OIFs was to determine the nexthop during
path. It is far better to get the OIF during path than for all routes
during route get.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>