Path: Handle recursive route with new route show
Since we stripped off getting fully resolved routes in routes.py,
we need to recursively resolve routes in path, which is really the
only place we need them (if we ever need them in topology, we may
have a decent refactoring effort, but thats another story). This patch
adds this logic to path. It also has a bug fix and an unrelated
enhancement.

The main observations about this logic include:
  * Multiple levels of recursion are supported, though we fix it
    to 2 now.
  * We automatically handle the recursed route resulting in an underlay
    lookup.
  * panos data is the only one in our setup that has recursive route
    test because only panos provides the relevant data.

The unrelated bug fix is to properly handle the case of whether a route
lookup is L2 or L3 when the first hop is the first hop switch, not the
endpoint of the address itself.

The unrelated enhancement is to allow path to work even when we neither
have the destination being a polled node nor have it in the ARP table of
the last hop switch. The unrelated consequence of this is that we'll
take a path all the way to the last hop switch. This enables us to
support tracing paths to service anycast addresses.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>