Improve EVPN VNI assert
EVPN VNI assert was slow because it attempted to verify reachability to
every remote VTEP known to a VTEP. This was done because once upon a
time one of the implementations didn't handle checking this reachability
before installing the routes.

We also enabled the check needed on Linux/FRR-based NOS (Cumulus &
SONiC) to check that the mac table has all the appropriate entries for
doing ingress replication.

Also for systems such as Junos that don't have a VXLAN interface, do not
do any Vxlan interface checks.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>