CLI: Add debug mode to display detailed stack trace
When we're running the CLI, sometimes we run into exceptions due to
bugs. Because we catch all exceptions and display only the error
message, it can be hard to troubleshoot where the real bug is.

This patch adds a debug mode to the context command set (set/clear) that
enables us to get the detailed exception when debug mode is enabled. By
default, the debug mode is off.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>