Route & Path improvements (#584)
* sq-poller: gracefully terminate with -i mode

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* Rename run_once internal variable to single_run_mode

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* Fix tests according to 'single_run_mode' variable name

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* Move describe table in xplore page

Signed-off-by: LucaNicosia <lucassa5@live.com>

* Pylint fixes

Signed-off-by: LucaNicosia <lucassa5@live.com>

* use describe function

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* sq-poller: Show input-dir and run-once in debug mode help message

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* Poller worker minor fixes (#580)

* sq-poller: allow to pass additional args to the inventory init function

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: rename Poller class to Worker

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: Move inventory intialization after service manager

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: make clearer the no nodes or no services messages

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* CLI: Add debug mode to display detailed stack trace (#586)

When we're running the CLI, sometimes we run into exceptions due to
bugs. Because we catch all exceptions and display only the error
message, it can be hard to troubleshoot where the real bug is.

This patch adds a debug mode to the context command set (set/clear) that
enables us to get the detailed exception when debug mode is enabled. By
default, the debug mode is off.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Add PR template (#578)

* add PR template

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* add PR template pt 2

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* Add contribute page

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* update contribute page

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* Move all test data into a single dir (#594)

This allows us to test remote data via REST. Without this, we'll have to
repeat the same tests for remote and local data which is a waste.
Furthermore, this allows us to test the REST API in a more consistent
way as well.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Routes: Get rid of populating oifs

Not every NOS provides the outgoing interface associated with a nexthop
for a given route. We attempted to normalize this by trying to fill in
the NOS by doing recursive route lookups. At scale, this blows up very
quickly.

The main reason for adding the OIFs was to determine the nexthop during
path. It is far better to get the OIF during path than for all routes
during route get.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Path: Handle recursive route with new route show

Since we stripped off getting fully resolved routes in routes.py,
we need to recursively resolve routes in path, which is really the
only place we need them (if we ever need them in topology, we may
have a decent refactoring effort, but thats another story). This patch
adds this logic to path. It also has a bug fix and an unrelated
enhancement.

The main observations about this logic include:
  * Multiple levels of recursion are supported, though we fix it
    to 2 now.
  * We automatically handle the recursed route resulting in an underlay
    lookup.
  * panos data is the only one in our setup that has recursive route
    test because only panos provides the relevant data.

The unrelated bug fix is to properly handle the case of whether a route
lookup is L2 or L3 when the first hop is the first hop switch, not the
endpoint of the address itself.

The unrelated enhancement is to allow path to work even when we neither
have the destination being a polled node nor have it in the ARP table of
the last hop switch. The unrelated consequence of this is that we'll
take a path all the way to the last hop switch. This enables us to
support tracing paths to service anycast addresses.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Test fixup for the route oif computation fix

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* flake8 fix for routes

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

Co-authored-by: Claudio Lorina <claudio.lorina@gmail.com>
Co-authored-by: LucaNicosia <lucassa5@live.com>
Co-authored-by: LucaNicosia <luca.nicosia@stardustsystems.net>
Co-authored-by: LucaNicosia <66721906+LucaNicosia@users.noreply.github.com>