Routes: always get prefix
When we're returning columns without a prefix, after the fix for only
looking for augmented fields the user specified, prefix no longer gets
returned. On Linux machines where the routes includes entries such as
127.0.0.0/8, which we filter out. But without returning prefix, we end
can't filter these entries and so we end up returning different results
for unique, show and show with columns that doesn't include the prefix.

This patch fixes that. It also cleans up some redundant code.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>