Schema: Support for nested augmented fields
engineobj: update api to get augmented fields

In engineobj.py, we returned dependent fields for every augmented field
in a table whether the augmented field was requested or not. Fixed that.

As a consequence, address.py and topology.py needed additional column
specifications to work correctly

In engineobj, we filter the dataframe returning only rows with matching
hostname and/or active status based on user specifications. When we do
that, we don't reset the index. Without this, when we attempt to update
the routes dataframe, we trigger the pandas changing values on a
section of the dataframe.

Fix this by also returning a dataframe with renumbered indices.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>