test_coalescer fix to handle not reading aug fields
When reading and writing dataframe in the coalescer tests, if we read
augmented fields as well (via columns=*), then attempting to write back
the data--the way we've coded write back in this test--fails because the
augmented fields are still part of the dataframe. To avoid this, read
only non-augmented columns from table.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>