parquetdb: Return all fields in current schema if requested
One of the beenfits of schema evolution support is the ability to read
older DB which may not contain the relevant fields in the DB and return
them as blanks. However, the code was not handling this part. It only
returned the fields that were retrieved from the data. This patch
returns all the requested fields in the current schema, blanking their
values if missing.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>