Fix ssh connection leakage
When resuming an ssh connection after a node has been unreachable for more than ~16 mins, multiple connections were attempted instead of one. That happened because the set() operation on the ssh_ready Event would resume all the tasks that were waiting for it. Replacing the Event with a Lock ensures that only the first task that started waiting on that lock will be resumed.

Signed-off-by: Claudio Usai <claudio.usai@stardustsystems.net>