Fix default in sqPollerCmd
The filter pollExcdPeriodCount needs to be set empty as default, otherwise the user cannot see the overall that but only the polling event that either have exceeded or not. In cases such as having an isolated device that would result in a timeout of the polling, these failing event would not be shown.

Signed-off-by: Claudio Usai <claudio.usai@stardustsystems.net>