LLDP: Add ability to return bond interfaces instead of mbr ports
IN many occasions, its required for LLDP to resolve the interface
names to their port-channel (or bond) names instead of the individual
member ports. This patch adds an option to do so.

When use_bond is "True", we resolve the individual member ports to their
owning port channel member name.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>