test_parsing_ospf: Fixup required after OSPF analyzer changes
We switched to relying on LLDP first to build out the peerHostname and
peerIfname columns in the returned ospf dataframe. However, this meant
that ipAddress field is not populated correctly for FRR-based NOS such
as Cumulus and SONiC. This is because FRR doesn't return that field in
its show command output.

So, this patch fixes the test_parsing_ospf test to expect ipAddress
field to be non-null for all NOS except Cumulus and SONiC.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>