Filter afisafi (#617)
* Add AFI-SAFI filter to BGP and topology

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* Tests to check afiSafi

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

* sqpoller: Change default status value to blank to show all values

The default value of the status filter was set to "all" to indicate that
without any specification, the user should be shown all rows regardless
of status. However, this caused the describe function to complain that
the user had provided an additional specification. sqpoller's status was
the only field with such a default. This patch makes the default value
of status the empty string.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>