IOSXE interfaces poller: Ignore invalid interfaces
Due to whatever reason, in some installations, IOSXE interfaces returns
entries with no type. This can happen if a transient interface such as
NDE_0 gets created between the time we run show interface and show
ip/ipv6 interface. Such interfaces only show up in the second command,
not the first. In such a situation, we fail to handle the ill-formed
entry.

The fix is to drop such malformed entries. If the interface is not
ephemeral, we'll catch it in the next run.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>