Xplore GUI: Handle view changes with time correctly
When the user specifies start and end time, we automatically
switch the view to all to display all the changes between specified
start and end times. However, when the user removes one or both of the
start and end time specifications, we do not revert the view to latest.
This is important to avoid trying to fetch and display too much data. A
simple removal of time filter shouldn't cause this problem.

This patch addresses this issue by reverting the view to latest whenever
we switch times. This does have the adverse behavior of switching the
view when the user had forcibly changed the view to latest, but is
playing with changing only the end or start times without specifying the
other. But, we'll mark this behavior for now and not attempt to address
it.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>