Xplore GUI: Limit rows shown at a time
Trying to display tables larger than 50K can typically take either a
very long time or fail with too much memory required to display the
info. To avoid this issue, we add support for start and end rows to
allow users to view the data in batches.

This model obviously breaks instant filters, if the data resides outside
the boundary of what is being viewed. To handle this, either reduce the
window of time in which you're viewing the data, qualify with namespace
or use pandas query string to filter.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>