sq-poller: ignore derived columns when comparing new and old data
When we retrieve the data from parquet, now we also return the derived
columns as empty fields, which will be later filled by the engine. This
caused a problem to poller which expected that fields in the newly
retrieved data. This patch tells the poller to compare only the
columns that are present in the newly retrieved records.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>