Fix show view=all setting a conditional default value for active_only
active_only is a field telling the engine whether to retrieve only the
active records (discarding everything has been deleted). If view=all,
by default, we would like to show all the records, in all the other
cases we want to see only the active records. This patch set the
active_only default value according to this logic.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>