Avoid crashes caused by passing very long hostnames or ifnames
In many situations, we're trying to limit the amount of information we
pull by providing the list of hostnames culled from another dataframe.
When the list of hosts culled this way is too big, pandas crashes (seen
this with > 500 hosts, for example). To avoid this, we do not pass
hostnames or ifnames created this way, but instead resort to merge to
get the data we need.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
(cherry picked from commit 816b2d8b69663a1661ddd03705bf638c3dd6a18c)