Handle broken filters and the ! operator
We weren't correctly handling hostname specification with "!". For
example, we returned an empty dataframe for lldp show
hostname="!leaf01". This patch addresses that. In the process of doing
that, we also fixed adding a ! operator with the regex operator. So, if
you invoke hostname='~!spine.*' we'll return a result that doesn't
include any hostname matching spine.* reogexp.

Furthermore, when the user specified lldp show hostname='!leaf01
!leaf02', we'd end up showing everything because we treat the space
separated list of hostnames as OR. However, with the use of the !
operator, it should become AND, not OR. This problem percolates
everywhere: hostname filter in engineobj.py, build_query_str in
shared/utils.py and namespace filter in db/parquet/parquetdb.py.

This was needed as we attempted to cleanup the use
hostname.unique().tolist().

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
(cherry picked from commit 8036abdffe331bafe24cd3377002e7440d9057a9)