CLI: Remove aliases as framework handles those aliases automatically
To prevent user confusion and frustration, we had added aliases to
various commands. Thus users could type sqpoller instead of sqPoller or
routes and route instead of only route, and so on. All this was
unnecessary and cluttered help and actually added to the confusion. The
CLI framework autocorrects and handles these cases correctly already,
and so remove the aliases.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>