network: fix displaying NS with filters
When users specified filters such as os, model or version, we displayed
namespaces which didn't pass those filters. The reason was how we did
the merge. These filters are available only to device table, not to
sqpoller table. As a consequence, we added back all the namespaces that
we had excluded when we did the merge of device table with poller table,
because we merged the device table into the poller table. If we had done
the opposite i.e. devdf.merge(pollertable...), this problem would not
have occurred. This patch adds that fix.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>