Various fixes for hangs/timeouts
This patch contains a set of fixes for IOSXE when privilege escalation
is required. We were resulting in irrecoverable hangs, timeouts, and so
on. They were caused by not handling erroneous returns correctly in some
cases.

This doesn't handle the case of avoiding lockout by trying the wrong
password too many times because we initiate connection over and over
again.

This patch fixes lock releases in case of certain errors, and finally,
we handle authentication failures to avoid locking out users by retrying
failed logins with the same bad password.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>