DB: Fix time handling
The DB is optimized to provide constant query times independent of the
time window specified. The code to read the set of files given the
variety of user inputs was buggy and read a lot more files than
necessary. This reduced the time to respond to user queries. This patch
addresses this.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>