Interfaces: Display the access port VLAN info correctly
Depending on the NOS, the interfaces output gathered doesn't always
provide the right port VLAN info (access ports or PVID for trunk ports).
However, if the user requests portmode, the VLAN info retrieved is
correct. To avoid this inconsistency and to show the correct VLAN info,
this patch assumes that we're fetching the portmode info if either the
portmode is specified or the vlan is specified.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>