Junos Interfaces: Parse VLAN and internal interfaces
We were incorrectly mapping lo.16385 and such as having a VLAN
of 16385 which is of course nonsense. If the number is > 4095,
we now make the VLAN 0. There are cases where we have interfaces such as
irb.4192 which get their VLAN initialized to 0. Am not handling those
correctly yet.

Internal interfaces in Junos like jsrv had their logical interfaces such
as jsrv.1 classified as non-internal interfaces which caused some issues
with assert and some inconsistent behavior. Now we make logical
interfaces of internal interfaces also internal, except for irb.<vlanid>
which gets classified as a VLAN (SVI).

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>