Node: Make device recognition simpler
To recognize what type of device a node was, we attempted to run
multiple commands on a node. Some of the commands--such as Cumulus cat
/etc/os-relase OR even NXOS show hostname--caused some of the devices,
especially IOSXE devices, to shutdown the SSH session.

This patch fixes that by running only one command--we pick show
version--as the first since non-IOS devices do not shutdown a session if
an invalid command is provided.

The patch involved handling the case of EOS which doesn't supply the
hostname in show version unlike all the other NOS we support. As a
consequence, we had to special case the handling of EOS devices.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>