Interfaces: Some misc VLAN fixes
In certain scenarios when we're trying to retrieve PVID and such, there
were a couple of bugs in the code causing crashes in the GUI status
display and in Path. These issues didn't exist in 0.17.0, This patch
addresses the fixes for those two issues. One was caused by an
indentation error, and the other because VLAN had an empty string
field. For the latter, we ensure that the confutils routines return vlan
as an integer.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>