Poller: More minor fixes and debug messages
We were continuing to attempt to poll when we had authentication
failures and other such errors. When we retrieved netbox inventory, it
was hard to know how many entries we got and how many we ignored because
of an error. Finally, when we raised NotImplementedError, we didn't trap
it in an exception handling clause causing an entire worker to die and
we didn't print additional info about what was not implemented and on
what device IP.

This patch addresses all this.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>