Tests: Update data update tests to handle new test data format
All test parquet data are stored under tests/data/parquet now instead of
being broken up further into subdirs belonging to different namespaces.
The main reason for this change was to enable testing of remote DB use
as well. However, with this change, we had not fixed the test update
code that enabled us to update or add new test data. This patch fixes
that.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>