Node: One more fixup after refactoring
We had the true/false signs mixed up in checking for a live
connection in case of IOS/IOSXE devices and that screwed up the
connectivity to those devices. Furthermore, if the user supplies a
device type, we need to switch to it first as the device might require
privilege escalation and if we don't switch first, we'll never manage to
connect to the device.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>