sq-poller: fix discovery attempts
During the node initialization, if we are not able to discover the node
type, we reschedule the discovery (i.e. unable to connect). However,
if we fail the discovery again we do not update the next attempt time,
so from this moment, for each new command issued on the node we perform
an attempt. This commit fixes this problem updating the time for each
attempt.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>